Restructure Footer
